+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Additional tests for listing files/versions
 * Ensured that changelog validation only happens on pull requests
 * Upgraded GitHub actions checkout to v3, python-setup to v4
+* Add `BucketTrackingMixin` class for tracking buckets created by `B2Api` instances
 
 ## [1.18.0] - 2022-09-20